New memory handling via range-checking/propagation functions
Replace direct access to large memory buffers for CPU and PPU memory with read/write functions that check address ranges and propagate to smaller buffers that represent the actual hardware. This simplifies the implementation of memory access side-effects (register shifts, etc.), and will make it much easier to implement mappers.