Eliminate audio stuttering
Now using a float to count elapsed cpu cycles to add audio samples to get as close as possible to the output sampling rate. Also increased the size of the circular buffer to avoid underflow. This seems to have fixed the problem of write samples being dropped without introducing much delay.