Fix save states for roms that use non-battery-backed PRG-RAM (like SMB3)
Changed the way the rom header is parsed to support detection of archaic iNes, iNes, and Nes 2.0 header. From this, I now correctly compute PRG-RAM size, and use this value to determine how much of it to serialize for save states. This fixes save states (and rewind by extension) for these roms.